Move string extensions to separate file
