+++ ProximityShare/Settings/Views/Components/DisplayPicture.swift
@@ -7,17 +7,6 @@
 
 import SwiftUI
 
-extension String {
-    func initials() -> String {
-        let words = self.components(separatedBy: " ")
-        var initials = words.compactMap { $0.first }.map { String($0) }.joined()
-        if initials.count > 2 {
-            initials = String(initials.first!) + String(initials.last!)
-        }
-        return initials.uppercased()
-    }
-}
-
 struct DisplayPicture: View {
     var name: String
     var size: CGFloat
